backend/requirements.txt
flask
flask-cors
gunicorn
sentence-transformers
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
numpy
openai-whisper
imageio-ffmpeg
requests
youtube-transcript-api